Move useMarketAndOffer, useOfferMode, useSyntheticOffer to files
