+++ src/components/PlaceOfferSection/hooks/useSyntheticOffer.ts
@@ -1,0 +1,24 @@
+import { useMemo } from 'react'
+
+import { useWallet } from '@solana/wallet-adapter-react'
+
+import { createEmptySyntheticOffer, useSyntheticOffers } from '@banx/store'
+
+export const useSyntheticOffer = (offerPubkey: string, marketPubkey: string) => {
+  const { publicKey: walletPubkey } = useWallet()
+
+  const { findOfferByPubkey, setOffer: setSyntheticOffer, removeOffer } = useSyntheticOffers()
+
+  const syntheticOffer = useMemo(() => {
+    return (
+      findOfferByPubkey(offerPubkey) ||
+      createEmptySyntheticOffer({ marketPubkey, walletPubkey: walletPubkey?.toBase58() || '' })
+    )
+  }, [findOfferByPubkey, marketPubkey, walletPubkey, offerPubkey])
+
+  const removeSyntheticOffer = () => {
+    removeOffer(syntheticOffer.marketPubkey)
+  }
+
+  return { syntheticOffer, removeSyntheticOffer, setSyntheticOffer }
+}
